extract Pagination functionality to the separate reducer
