+++ src/actions/PaginationAction.js
@@ -1,0 +1,6 @@
+import * as types from 'constants/actionTypes/PaginationActionTypes';
+
+export const setPage = (currentPage) => ({
+  type: types.SET_PAGE,
+  currentPage
+});
