+++ src/constants/actionTypes/PaginationActionTypes.js
@@ -1,0 +1,1 @@
+export const SET_PAGE = 'SET_PAGE';
