+++ src/constants/actionTypes/PostsActionTypes.js
@@ -2,4 +2,4 @@ export const FETCH_POSTS_REQUEST = 'FETCH_POSTS_REQUESTS';
 export const FETCH_POSTS_SUCCESS = 'FETCH_POSTS_SUCCESS';
 export const FETCH_POSTS_ERROR = 'FETCH_POSTS_ERROR';
 export const ADD_LIKE = 'ADD_LIKE';
-export const SET_PAGE = 'SET_PAGE';
+// export const SET_PAGE = 'SET_PAGE';
